perf(ui): convert /login to server component
